Removed the /internal/ directory in Keyvault Keys and move headers to private that don't need to be public. (#2625)
* Removed the /internal/ directory in Keyvault Keys and move headers to
private that don't need to be public.

* Clang format.

* Move CryptographyProvider and RCC into private headers as well.

* Fix clang formatting.
