+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/rsa_cryptography_provider.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/keys/internal/cryptography/rsa_cryptography_provider.hpp"
+#include "../private/rsa_cryptography_provider.hpp"
 
 #include <memory>
 #include <string>
