+++ sdk/keyvault/azure-security-keyvault-keys/src/private/local_cryptography_provider_factory.hpp
@@ -9,9 +9,10 @@
 
 #pragma once
 
-#include "azure/keyvault/keys/internal/cryptography/cryptography_provider.hpp"
-#include "azure/keyvault/keys/internal/cryptography/rsa_cryptography_provider.hpp"
 #include "azure/keyvault/keys/key_vault_key.hpp"
+
+#include "cryptography_provider.hpp"
+#include "rsa_cryptography_provider.hpp"
 
 #include <memory>
 #include <string>
